refactor: Añadiendo funcionesHTML
Añadiendo un nuevo documento para que este mejor estructurado
